moving parser stuff to non adapter specific location
