--- .run/vectorsearch-consistency-levels.run.xml
@@ -1,15 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="vectorsearch-consistency-levels" type="JarApplication">
-    <extension name="software.aws.toolkits.jetbrains.core.execution.JavaAwsConnectionExtension">
-      <option name="credential" />
-      <option name="region" />
-      <option name="useCurrentConnection" value="false" />
-    </extension>
-    <option name="JAR_PATH" value="$PROJECT_DIR$/nb5/target/nb5.jar" />
-    <option name="PROGRAM_PARAMETERS" value="vector-search.yaml reads errors=stop driverconfig=driver-config.json dimensions=128 read_ratio=1 main-cycles=1 keyspace=baselines128 --report-csv-to metrics read_cl=LOCAL_ONE -v --show-stacktraces" />
-    <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/local/vectors-consistency" />
-    <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="true" />
-    <option name="ALTERNATIVE_JRE_PATH" value="17" />
-    <method v="2" />
-  </configuration>
-</component>
