+++ virtdata-lib-hdf5/src/main/java/io/nosqlbench/virtdata/library/hdf5/from_long/to_string/predicate_parser/JAWDatasetParser.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 nosqlbench
+ * Copyright (c) 2023-2024 nosqlbench
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,17 +12,26 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-package io.nosqlbench.datamappers.functions.hdf_to_cql;
+package io.nosqlbench.virtdata.library.hdf5.from_long.to_string.predicate_parser;
+
+import io.nosqlbench.virtdata.library.hdf5.from_long.to_string.predicate_parser.DatasetParser;
+
 /**
  * This class is used to parse the raw JSON from the HDF dataset into a CQL predicate. This implementation
- * accepts a string consisting of the desired CQL predicate as translated from the original jsonl files and
- * simply returns the raw string, hence the name NoopDatasetParser.
+ * accepts a string consisting of the desired CQL predicate as translated from the original jsonl files
+ * and simply adds the WHERE keyword to the beginning of the string if it is not already present, hence
+ * the new Just Add Where (JAW) parser.
  */
-public class NoopDatasetParser implements DatasetParser {
+public class JAWDatasetParser implements DatasetParser {
+    private static final String WHERE = "WHERE";
     @Override
     public String parse(String raw) {
+        if (!raw.toUpperCase().startsWith(WHERE)) {
+            raw = WHERE + " " + raw;
+        }
         return raw;
     }
 }
